requirements.txt
django-easy-pdf
django-crispy-forms